Website can now take display scores instead of heat scores
Display scores can now be used, instead of heat scores, as input for the
website. The scores can be negative numbers. If the file is given, the
display scores are labeled as heat and processed in the same manner as
heat scores by subnetworks.html

 - Added the optional argument "-dsf/--display_score_file" to the simpleRun.py and
makeResultsWebsite.py parsers.
 - Added function "load_display_score_tsv" which, like "load_heat_tsv", load the TSV as
a dictionary for JSON encoding.
 - Modified get_nodes and get_component_json to accomodate the extra parameter "d_score"
(display score).